# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Documentação das API do sistema Jobify."
  version: "1.0.0"
  title: "Jobify API"
  contact:
    email: "bruno95edu@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/brunoedu/Jobify/1.0.0"
...
schemes:
- "https"
paths:
  /categories/jobs:
    get:
      tags:
      - "admins"
      - "users"
      summary: "Lista todas as categorias de serviço"
      description: "Lista todas as categorias de serviços podendo conter alguns filtros\n"
      operationId: "getAllJobCategories"
      produces:
      - "application/json"
      parameters:
      - name: "skip"
        in: "query"
        description: "Número de categorias a pular para paginação"
        required: false
        type: "integer"
        minimum: 0
        format: "int32"
      - name: "limit"
        in: "query"
        description: "Número máximo de categorias a retornar"
        required: false
        type: "integer"
        maximum: 50
        minimum: 0
        format: "int32"
      responses:
        200:
          description: "Lista de categorias de serviço"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/JobCategory"
        400:
          description: "Parâmetros inválidos"
        401:
          description: "Não autorizado"
        403:
          description: "Não permitido"
        500:
          description: "Erro interno de servidor"
      security:
      - api_key: []
      x-swagger-router-controller: "Admins"
    post:
      tags:
      - "admins"
      summary: "Cria uma nova categoria de serviço"
      description: "Cria uma nova categoria de serviço"
      operationId: "postJobCategory"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "jobCategory"
        description: "Categoria de serviço a ser adicionada"
        required: true
        schema:
          $ref: "#/definitions/JobCategory"
      responses:
        201:
          description: "Categoria criada"
        400:
          description: "Objeto de categoria inválido"
        401:
          description: "Não autorizado"
        403:
          description: "Não permitido"
        409:
          description: "Uma categoria com este nome já existe"
        500:
          description: "Erro interno de servidor"
      security:
      - api_key: []
      x-swagger-router-controller: "Admins"
    put:
      tags:
      - "admins"
      summary: "Edita uma nova categoria de serviço"
      description: "Edita uma nova categoria de serviço"
      operationId: "putJobCategory"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "jobCategory"
        description: "Categoria de serviço a ser editada"
        required: true
        schema:
          $ref: "#/definitions/JobCategory"
      responses:
        200:
          description: "Categoria editada"
        400:
          description: "Objeto de categoria inválido"
        401:
          description: "Não autorizado"
        403:
          description: "Não permitido"
        404:
          description: "Categoria não encontrada"
        500:
          description: "Erro interno de servidor"
      security:
      - api_key: []
      x-swagger-router-controller: "Admins"
  /categories/suggestions:
    get:
      tags:
      - "admins"
      - "users"
      summary: "Lista todas as categorias de sugestões"
      description: "Lista todas as categorias de sugestões podendo conter alguns filtros\n"
      operationId: "getAllSuggestionsCategories\L"
      produces:
      - "application/json"
      parameters:
      - name: "skip"
        in: "query"
        description: "Número de categorias a pular para paginação"
        required: false
        type: "integer"
        minimum: 0
        format: "int32"
      - name: "limit"
        in: "query"
        description: "Número máximo de categorias a retornar"
        required: false
        type: "integer"
        maximum: 50
        minimum: 0
        format: "int32"
      responses:
        200:
          description: "Lista de categorias de sugestões"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SuggestionCategory"
        400:
          description: "Parâmetros inválidos"
        401:
          description: "Não autorizado"
        403:
          description: "Não permitido"
        500:
          description: "Erro interno de servidor"
      security:
      - api_key: []
      x-swagger-router-controller: "Admins"
    post:
      tags:
      - "admins"
      summary: "Cria uma nova categoria de sugestão"
      description: "Cria uma nova categoria de sugestão"
      operationId: "postSuggestionCategory"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "suggestionCategory"
        description: "Categoria de sugestão a ser adicionada"
        required: true
        schema:
          $ref: "#/definitions/SuggestionCategory"
      responses:
        201:
          description: "Categoria criada"
        400:
          description: "Objeto de categoria inválido"
        401:
          description: "Não autorizado"
        403:
          description: "Não permitido"
        409:
          description: "Uma categoria com este nome já existe"
        500:
          description: "Erro interno de servidor"
      security:
      - api_key: []
      x-swagger-router-controller: "Admins"
    put:
      tags:
      - "admins"
      summary: "Edita uma nova categoria de sugestão"
      description: "Edita uma nova categoria de sugestão"
      operationId: "putSuggestionCategory"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "suggestionCategory"
        description: "Categoria de sugestão a ser editada"
        required: true
        schema:
          $ref: "#/definitions/SuggestionCategory"
      responses:
        200:
          description: "Categoria editada"
        400:
          description: "Objeto de categoria inválido"
        401:
          description: "Não autorizado"
        403:
          description: "Não permitido"
        404:
          description: "Categoria não encontrada"
        500:
          description: "Erro interno de servidor"
      security:
      - api_key: []
      x-swagger-router-controller: "Admins"
  /suggestions:
    get:
      tags:
      - "admins"
      summary: "Lista todas as sugestões deixadas por usuários"
      description: "Lista todas as sugestões de usuários podendo conter alguns filtros\n"
      operationId: "getAllSuggestions"
      produces:
      - "application/json"
      parameters:
      - name: "skip"
        in: "query"
        description: "Número de sugestões a pular para paginação"
        required: false
        type: "integer"
        minimum: 0
        format: "int32"
      - name: "limit"
        in: "query"
        description: "Número máximo de sugestões a retornar"
        required: false
        type: "integer"
        maximum: 50
        minimum: 0
        format: "int32"
      responses:
        200:
          description: "Lista de sugestões"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Suggestion"
        400:
          description: "Parâmetros inválidos"
        401:
          description: "Não autorizado"
        403:
          description: "Não permitido"
        500:
          description: "Erro interno de servidor"
      security:
      - api_key: []
      x-swagger-router-controller: "Admins"
...
  /authenticate:
    post:
      tags:
      - "admins"
      - "users"
      summary: "Autentica um usuário no sistema liberando seu token de acesso"
      description: "Autnetica um usuário no sistema liberando seu token de acesso"
      operationId: "authUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "credentials"
        description: "Dados de autenticação de usuário"
        required: true
        schema:
          $ref: "#/definitions/Auth"
      responses:
        200:
          description: "Usuário autenticado"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Objeto de autenticação inválido"
        404:
          description: "Email não encontrado"
        500:
          description: "Erro interno de servidor"
      x-swagger-router-controller: "Admins"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "x-access-token"
    in: "header"
definitions:
...
  Auth:
    type: "object"
    required:
    - "email"
    - "password"
    properties:
      email:
        type: "string"
        example: "bruno@email.com"
        description: "Email do usuário"
        minLength: 3
        maxLength: 100
      password:
        type: "string"
        example: "Senh@123"
        minLength: 8
        maxLength: 100
    example:
      password: "Senh@123"
      email: "bruno@email.com"
  User:
    type: "object"
    required:
    - "_id"
    - "cellphone"
    - "email"
    - "name"
    - "password"
    properties:
      _id:
        type: "string"
        format: "objectId"
        example: "507f191e810c19729de860ea"
        description: "Id do usuário"
      name:
        type: "string"
        example: "Bruno"
        description: "Nome do usuário"
        minLength: 3
        maxLength: 100
      email:
        type: "string"
        example: "bruno@email.com"
        description: "Email do usuário"
        minLength: 3
        maxLength: 100
      password:
        type: "string"
        example: "Senh@123"
        minLength: 8
        maxLength: 100
      cellphone:
        type: "number"
        example: 1.2999999999E10
        description: "Celular do usuário"
      picture:
        type: "string"
        format: "url"
        example: "https://res.cloudinary.com/brunoedu/image/upload/v1508965936/sample.jpg"
        description: "Foto do usuário"
      lat:
        type: "number"
        example: 40.7143528
        description: "Latitude do usuário capturada de tempos em tempos para cálculo\
          \ das distâncias nos filtros de busca"
      lng:
        type: "number"
        example: -74.0059731
        description: "Longitude do usuário capturada de tempos em tempos para cálculo\
          \ das distâncias nos filtros de busca"
      push_web_token:
        type: "string"
        format: "guid"
        example: "50a5ecd2-bebc-440f-9486-bd17b9503a21"
        description: "Token do OneSignal para envio de push notifications web"
      push_mobile_token:
        type: "string"
        format: "guid"
        example: "50a5ecd2-bebc-440f-9486-bd17b9503a21"
        description: "Token do OneSignal para envio de push notifications mobile"
      socket_token:
        type: "string"
        example: "8411473621394413000"
        description: "Token do Socket.io para comunicação via chat"
      push_web_enabled:
        type: "boolean"
        example: true
        description: "Configuração de recebimento de notificações push na web"
      push_mobile_enabled:
        type: "boolean"
        example: true
        description: "Configuração de recebimento de notificações push mobile"
      email_enabled:
        type: "boolean"
        example: true
        description: "Configuração de recebimento de emails"
      tutorial_completed:
        type: "boolean"
        example: true
        description: "Verificação ser usuário já passou pelo tutoprial"
      categories:
        type: "array"
        description: "Lista de categorias de serviço"
        items:
          $ref: "#/definitions/JobCategoryUser"
      type:
        type: "string"
        example: "user"
        description: "Tipo de usuário (user/admin)"
        enum:
        - "user"
        - "admin"
      rating:
        type: "number"
        example: 4.5
        description: "Média de avaliações do usuário"
      jobs_accepted:
        type: "number"
        example: 10.0
        description: "Serviços aceitados"
      jobs_requested:
        type: "number"
        example: 2.0
        description: "Serviços solicitados"
    example:
      lng: -74.0059731
      email_enabled: true
      jobs_requested: 2.0
      rating: 4.5
      push_mobile_token: "50a5ecd2-bebc-440f-9486-bd17b9503a21"
      push_web_enabled: true
      tutorial_completed: true
      type: "user"
      picture: "https://res.cloudinary.com/brunoedu/image/upload/v1508965936/sample.jpg"
      password: "Senh@123"
      push_web_token: "50a5ecd2-bebc-440f-9486-bd17b9503a21"
      push_mobile_enabled: true
      jobs_accepted: 10.0
      name: "Bruno"
      cellphone: 1.2999999999E10
      socket_token: "8411473621394413000"
      _id: "507f191e810c19729de860ea"
      categories:
      - min_value: 150.0
        km: 30.0
        name: "Encanador"
        icon: "https://res.cloudinary.com/brunoedu/image/upload/v1508965936/sample.jpg"
        description: "Você conserta sistemas hidráulicos"
        _id: "507f191e810c19729de860ea"
      - min_value: 150.0
        km: 30.0
        name: "Encanador"
        icon: "https://res.cloudinary.com/brunoedu/image/upload/v1508965936/sample.jpg"
        description: "Você conserta sistemas hidráulicos"
        _id: "507f191e810c19729de860ea"
      email: "bruno@email.com"
      lat: 40.7143528
  JobCategory:
    type: "object"
    required:
    - "_id"
    - "description"
    - "icon"
    - "name"
    properties:
      _id:
        type: "string"
        format: "objectId"
        example: "507f191e810c19729de860ea"
        description: "Id da categoria de serviço"
      name:
        type: "string"
        example: "Encanador"
        description: "Nome da categoria de serviço"
        minLength: 3
        maxLength: 50
      description:
        type: "string"
        example: "Você conserta sistemas hidráulicos"
        description: "Descrição da categoria de serviço"
        minLength: 3
        maxLength: 256
      icon:
        type: "string"
        format: "url"
        example: "https://res.cloudinary.com/brunoedu/image/upload/v1508965936/sample.jpg"
        description: "Ícone da categoria"
    example:
      name: "Encanador"
      icon: "https://res.cloudinary.com/brunoedu/image/upload/v1508965936/sample.jpg"
      description: "Você conserta sistemas hidráulicos"
      _id: "507f191e810c19729de860ea"
  JobCategoryUser:
    type: "object"
    required:
    - "_id"
    - "description"
    - "icon"
    - "km"
    - "min_value"
    - "name"
    properties:
      _id:
        type: "string"
        format: "objectId"
        example: "507f191e810c19729de860ea"
        description: "Id da categoria de serviço"
      name:
        type: "string"
        example: "Encanador"
        description: "Nome da categoria de serviço"
        minLength: 3
        maxLength: 50
      description:
        type: "string"
        example: "Você conserta sistemas hidráulicos"
        description: "Descrição da categoria de serviço"
        minLength: 3
        maxLength: 256
      icon:
        type: "string"
        format: "url"
        example: "https://res.cloudinary.com/brunoedu/image/upload/v1508965936/sample.jpg"
        description: "Ícone da categoria"
      min_value:
        type: "number"
        example: 150.0
        description: "Valor mínimo cobrado pelo serviço"
      km:
        type: "number"
        example: 30.0
        description: "Km máxima de atendimento"
    example:
      min_value: 150.0
      km: 30.0
      name: "Encanador"
      icon: "https://res.cloudinary.com/brunoedu/image/upload/v1508965936/sample.jpg"
      description: "Você conserta sistemas hidráulicos"
      _id: "507f191e810c19729de860ea"
...
  SuggestionCategory:
    type: "object"
    required:
    - "_id"
    - "description"
    - "name"
    properties:
      _id:
        type: "string"
        format: "objectId"
        example: "507f191e810c19729de860ea"
        description: "Id da categoria de sugestão"
      name:
        type: "string"
        example: "Melhorias"
        description: "Nome da categoria de sugestão"
        minLength: 3
        maxLength: 50
      description:
        type: "string"
        example: "Descreva o que podemos melhorar para melhor atendê-lo"
        description: "Descrição da categoria de sugestão"
        minLength: 3
        maxLength: 256
    example:
      name: "Melhorias"
      description: "Descreva o que podemos melhorar para melhor atendê-lo"
      _id: "507f191e810c19729de860ea"
  Suggestion:
    type: "object"
    required:
    - "_id"
    - "description"
    - "suggestion_category"
    properties:
      _id:
        type: "string"
        format: "objectId"
        example: "507f191e810c19729de860ea"
        description: "Id da sugestão"
      suggestion_category:
        $ref: "#/definitions/SuggestionCategory"
      description:
        type: "string"
        example: "Olá, gostaria que o aplicativo carregasse mais rapidamente."
        description: "Texto da sugestão do usuário"
        minLength: 3
        maxLength: 256
    example:
      suggestion_category:
        name: "Melhorias"
        description: "Descreva o que podemos melhorar para melhor atendê-lo"
        _id: "507f191e810c19729de860ea"
      description: "Olá, gostaria que o aplicativo carregasse mais rapidamente."
      _id: "507f191e810c19729de860ea"
...